add receive auto address
