+++ assets/domain/src/main/java/com/easy/assets/domain/repository/AssetRepository.kt
@@ -6,10 +6,16 @@ import com.easy.core.consts.ChainId
 import java.math.BigInteger
 
 interface AssetRepository {
-    fun address(legacy: Boolean): String
-    suspend fun signTransaction(): String
+    fun address(slug: String, legacy: Boolean): String
+    suspend fun signTransaction(slug: String): String
     suspend fun balance(address: String, chainId: ChainId, contractAddress: String): BigInteger
-    suspend fun transactions(address: String, chainId: ChainId, offset: Int, limit: Int, contractAddress: String): Result<List<Transaction>>
+    suspend fun transactions(
+        address: String,
+        chainId: ChainId,
+        offset: Int,
+        limit: Int,
+        contractAddress: String
+    ): Result<List<Transaction>>
 
     suspend fun assets(): Result<Map<String, List<AssetInfo>>>
 }
