+++ assets/domain/src/main/java/com/easy/assets/domain/use_case/AssetsUseCases.kt
@@ -1,6 +1,7 @@
 package com.easy.assets.domain.use_case
 
 data class AssetsUseCases(
+    val address: CoinAddress,
     val assets: Assets,
     val balance: AssetBalance,
     val transactions: AssetTransactions
