+++ assets/presentation/src/main/java/com/easy/assets/presentation/routers/NavConfig.kt
@@ -10,8 +10,8 @@ import androidx.navigation.NavType
 import androidx.navigation.navArgument
 import com.easy.assets.presentation.detail.AssetDetailScreen
 import com.easy.assets.presentation.receive.ReceiveScreen
-import com.easy.assets.presentation.send.NormalSendAddressScreen
-import com.easy.assets.presentation.send.NormalSendAmountScreen
+import com.easy.assets.presentation.send.address.NormalSendAddressScreen
+import com.easy.assets.presentation.send.amount.NormalSendAmountScreen
 import com.google.accompanist.navigation.animation.composable
 import com.google.accompanist.navigation.animation.navigation
 
@@ -56,7 +56,12 @@ fun NavGraphBuilder.assetsGraph(navController: NavController) {
                 navController.navigate(nav.router())
             }
         }
-        composable(route = AssetRouter.ASSET_RECEIVE,
+        composable(
+            route = AssetRouter.ASSET_RECEIVE + "/{address}",
+            arguments = listOf(
+                navArgument("address") {
+                    type = NavType.StringType
+                }),
             enterTransition = {
                 fadeIn(animationSpec = tween(700))
             },
@@ -69,7 +74,8 @@ fun NavGraphBuilder.assetsGraph(navController: NavController) {
             popExitTransition = {
                 fadeOut(animationSpec = tween(700))
             }) {
-            ReceiveScreen(address = "0x6b175474e89094c44da98b954eedeac495271d0f") {
+            val address = it.arguments?.getString("address")!!
+            ReceiveScreen(address = address) {
                 navController.navigateUp()
             }
         }
