+++ assets/presentation/src/main/java/com/easy/assets/presentation/send/amount/SendAmountState.kt
@@ -1,0 +1,6 @@
+package com.easy.assets.presentation.send.amount
+
+data class SendAmountState(
+    val amount: String = "",
+    val memo: String = ""
+)
